Minor tweaks in preparation for SwiftGRPCNIO.
In detail:

- Fix an error in `.travis-install.sh`
- Improve the Carthage steps in the `Makefile`
- Move the `StatusCode` enum into a dedicated file.
- Re-run `make project-carthage` and commit a few missing schemes.
