+++ Makefile
@@ -43,10 +43,10 @@ xcodebuild: project
 		xcodebuild -project SwiftGRPC.xcodeproj -configuration "Debug" -parallelizeTargets -target SwiftGRPC -target Echo -target Simple -target protoc-gen-swiftgrpc build
 
 build-carthage:
-	carthage build --no-skip-current
+	carthage build -project SwiftGRPC-Carthage.xcodeproj --no-skip-current
 
 build-carthage-debug:
-	carthage build --no-skip-current --configuration Debug --platform iOS, macOS
+	carthage build -project SwiftGRPC-Carthage.xcodeproj --no-skip-current --configuration Debug --platform iOS, macOS
 
 clean:
 	-rm -rf Packages
