+++ SwiftGRPC-Carthage.xcodeproj/xcshareddata/xcschemes/Echo.xcscheme
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Scheme LastUpgradeVersion = "9999" version = "1.3">
+  <BuildAction parallelizeBuildables = "YES" buildImplicitDependencies = "YES">
+    <BuildActionEntries>
+      <BuildActionEntry buildForTesting = "YES" buildForRunning = "YES" buildForProfiling = "YES" buildForArchiving = "YES" buildForAnalyzing = "YES">
+        <BuildableReference
+          BuildableIdentifier = "primary"
+          BuildableName = "'$(TARGET_NAME)'"
+          BlueprintName = "Echo"
+          ReferencedContainer = "container:SwiftGRPC-Carthage.xcodeproj">
+        </BuildableReference>
+      </BuildActionEntry>
+    </BuildActionEntries>
+  </BuildAction>
+  <TestAction
+    buildConfiguration = "Debug"
+    selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
+    selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
+    shouldUseLaunchSchemeArgsEnv = "YES"
+    codeCoverageEnabled = "NO">
+    <Testables>
+        <TestableReference
+          skipped = "NO">
+          <BuildableReference
+            BuildableIdentifier = "primary"
+            BuildableName = "'$(TARGET_NAME)'"
+            BlueprintName = "SwiftGRPCTests"
+            ReferencedContainer = "container:SwiftGRPC-Carthage.xcodeproj">
+          </BuildableReference>
+        </TestableReference>
+    </Testables>
+  </TestAction>
+<LaunchAction
+   buildConfiguration = "Debug"
+   selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
+   selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
+   launchStyle = "0"
+   useCustomWorkingDirectory = "NO"
+   ignoresPersistentStateOnLaunch = "NO"
+   debugDocumentVersioning = "YES"
+   debugServiceExtension = "internal"
+   allowLocationSimulation = "YES">
+   <BuildableProductRunnable
+      runnableDebuggingMode = "0">
+      <BuildableReference
+         BuildableIdentifier = "primary"
+         BuildableName = "'$(TARGET_NAME)'"
+         BlueprintName = "Echo"
+         ReferencedContainer = "container:SwiftGRPC-Carthage.xcodeproj">
+      </BuildableReference>
+   </BuildableProductRunnable>
+   <AdditionalOptions>
+   </AdditionalOptions>
+</LaunchAction>
+</Scheme>
